refactor: implement SubtitleSection component with enhanced props handling and tracking functionality
